+++ src/pipeline/pipeline.ml
@@ -141,20 +141,29 @@ let parse_verification_file = parse_file' Lexer.mode_verification
 
 type resolve_result = (Syntax.prog * ResolveImport.resolved_imports) Diag.result
 
-let resolve_flags ~(enhanced_migration : string option) pkg_opt  =
+let resolve_flags ~is_main ~base pkg_opt =
+  let resolve_path_flag flag =
+    if not is_main then !flag else
+    !flag |> Option.map (fun p ->
+      if not (!Flags.ocaml_js && Filename.is_relative p) then p else
+      let base_dir = if Sys.is_directory base then base else Filename.dirname base in
+      (* moc.js has no CWD, so resolve relative flag paths against the source file's directory *)
+      let p = Filename.concat base_dir p |> Lib.FilePath.normalise in
+      flag := Some p; p)
+  in
   ResolveImport.{
     package_urls = !Flags.package_urls;
     actor_aliases = !Flags.actor_aliases;
-    actor_idl_path = !Flags.actor_idl_path;
+    actor_idl_path = resolve_path_flag Flags.actor_idl_path;
     include_all_libs = pkg_opt = None && Flags.(!all_libs || !ai_errors || Option.is_some !implicit_package);
-    enhanced_migration
+    enhanced_migration = if is_main then resolve_path_flag Flags.enhanced_migration else None;
   }
 
 let resolve_prog (prog, base) : resolve_result =
   Diag.map
     (fun libs -> (prog, libs))
     (ResolveImport.resolve
-       (resolve_flags ~enhanced_migration:!Flags.enhanced_migration None)
+       (resolve_flags ~is_main:true ~base None)
        prog base)
 
 let resolve_progs =
@@ -461,7 +470,7 @@ let chase_imports_cached parsefn senv0 imports scopes_map
         let* prog, base = parsefn ri.Source.at f in
         let* () = Static.prog prog in
         let cur_pkg_opt = if lib_pkg_opt <> None then lib_pkg_opt else pkg_opt in
-        let* more_imports = ResolveImport.resolve (resolve_flags ~enhanced_migration:None cur_pkg_opt) prog base in
+        let* more_imports = ResolveImport.resolve (resolve_flags ~is_main:false ~base cur_pkg_opt) prog base in
         let* () = go_set cur_pkg_opt more_imports in
         let lib = lib_of_prog f prog in
         let* sscope = check_lib !senv cur_pkg_opt lib in